resume_parser_project/requirements.txt
pdfplumber
python-docx
pandas
openpyxl
streamlit
